Don't unconditionally free history atexit
The memory allocated for the history entries will be automatically freed
by the operating system on program termination. Manually freeing it in
an atexit(3) handler prevents the caller from using it in his/her atexit
handler.

I noticed this because I call linenoiseHistorySave() in my atexit
handler which doesn't work without this change. Nonetheless, it might be
handy to free memory on exit when debugging memory leaks with valgrind,
therefore the old behaviour can be restored by compiling with
`-DVALGRIND`.